backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "This is the backend for the Odessey project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
